few more updates in dfw pkg
