+++ pkg/synthesis/preProcessing.go
@@ -19,12 +19,12 @@ func preProcessing(categoriesSpecs []*dfw.CategorySpec) (categoryToPolicy map[co
 	groupToConjunctions := map[string][]*symbolicexpr.Conjunction{} // caching groups' Conjunctions
 	for _, category := range categoriesSpecs {
 		categoryPolicy := symbolicPolicy{}
-		if len(category.AllRules.Outbound)+len(category.AllRules.Inbound) == 0 {
+		if len(category.EvaluatedRules.Outbound)+len(category.EvaluatedRules.Inbound) == 0 {
 			continue
 		}
-		categoryPolicy.inbound = append(categoryPolicy.inbound, convertRulesToSymbolicPaths(category.AllRules.Inbound,
+		categoryPolicy.inbound = append(categoryPolicy.inbound, convertRulesToSymbolicPaths(category.EvaluatedRules.Inbound,
 			category.Category, groupToConjunctions)...)
-		categoryPolicy.outbound = append(categoryPolicy.outbound, convertRulesToSymbolicPaths(category.AllRules.Outbound,
+		categoryPolicy.outbound = append(categoryPolicy.outbound, convertRulesToSymbolicPaths(category.EvaluatedRules.Outbound,
 			category.Category, groupToConjunctions)...)
 
 		categoryToPolicy[category.Category] = &categoryPolicy
